feat: default values yaml
NOTE: this commit will slow down build time if values are not manually provided. this is because it looks in the newly created `defaults.yaml` file for default values instead of defining them in the code. i believe this is the best and only approach to providing default values, but i will continue to look for ways to regain compile speed